coded a function to convert rgb to hex and linked the color written in the file to the map

+++ save/includes/structs.h
@@ -8,9 +8,9 @@ typedef struct	s_graphics
 	char	*dir_WE;
 	char	*dir_EA;
 	int		*floor;
-	int		hex_floor;
+	unsigned long int		hex_floor;
 	int		*ceiling;
-	int		hex_ceiling;
+	unsigned long int		hex_ceiling;
 }				t_graphics;
 
 typedef struct	s_img
